Split resolver absence tests
